+++ src/components/Home/Authenticated/buttons/index.ts
@@ -1,3 +1,2 @@
 export { NewTripButton } from "./NewTripButton";
 export { JoinTripButton } from "./JoinTripButton";
-export { CopyTripIdButton } from "./CopyTripIdButton";
