+++ src/components/Trip/TripChat.tsx
@@ -1,0 +1,9 @@
+export const TripChat = () => {
+  return (
+    <div className="flex flex-col gap-3">
+      <label>send message</label>
+      <input />
+      <button type="button">send</button>
+    </div>
+  );
+};
