+++ src/components/Trip/buttons/ManageParticipants.tsx
@@ -1,11 +1,11 @@
 import { Transition } from "@headlessui/react";
 import * as Dialog from "@radix-ui/react-dialog";
-import { CrossIcon } from "../../icons";
+import { CrossIcon } from "../../../icons";
 import { clsx } from "clsx";
 import { type FC, type MouseEvent, Fragment, useState, useEffect } from "react";
 import { motion } from "framer-motion";
-import { regularFont } from "../../fonts";
-import { api } from "../../utils/api";
+import { regularFont } from "../../../fonts";
+import { api } from "../../../utils/api";
 import { useRouter } from "next/router";
 
 interface Props {
